requirements.txt
# requirements.txt

# --- 我必须锁定的核心包 ---
transformers==4.47.0
numpy==1.26.4

# --- 其他包，让 pip 自动选择兼容的版本 ---
torch
torchvision
huggingface-hub
sentencepiece
jinja2
pydantic
timm
tiktoken
accelerate
sentence_transformers
gradio
openai
einops
pillow
sse-starlette
bitsandbytes
spaces
